+++ Tasks/UnityBuild/UnityBuildV3/CHANGELOG.md
@@ -9,11 +9,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-- Unexpected token build fail issue (https://github.com/Dinomite-Studios/unity-azure-pipelines-tasks/issues/199)
+- Unexpected token build fail issue [GitHub Issue](https://github.com/Dinomite-Studios/unity-azure-pipelines-tasks/issues/199)
+- This task uses Node 6 execution handler, which will be deprecated soon [GitHub Issue](https://github.com/Dinomite-Studios/unity-azure-pipelines-tasks/issues/190)
 
 ### Changed
 
-- Use new shared library for common task features
+- Updated minimum required agent version to `2.144.0`
+- Use new shared library `@dinomite-studios/unity-azure-pipelines-tasks-lib` for common task features
 
 ## [3.1.12]